- backend/src/resolvers/Mutation - adding the logic to check if the user deleting an item has the permissions to do so - frontend/components/DeleteItem - adding an alert if an error occurs while trying to delete an item

+++ backend/src/resolvers/Mutation.js
@@ -38,7 +38,17 @@ const Mutations = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user{ id} }`);
+
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(
+      permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if(!ownsItem && !hasPermissions) {
+      throw new Error('You do not have permissions to delete this item.');
+    }
+
     return ctx.db.mutation.deleteItem({ where }, info);
   },
   async signup(parent, args, ctx, info) {
